TSK-1369: Refactor TaskHistoryCustomField from events package to events/task
