+++ lib/taskana-core/src/main/java/pro/taskana/spi/history/api/events/task/TaskHistoryCustomField.java
@@ -1,4 +1,4 @@
-package pro.taskana.spi.history.api.events;
+package pro.taskana.spi.history.api.events.task;
 
 public enum TaskHistoryCustomField {
   CUSTOM_1,
